add test for domain layer, fix minor data type inconsistencies after testing

+++ src/main/java/lv/klix/oas/domain/Offer.java
@@ -28,7 +28,7 @@ public class Offer extends BaseEntity {
 
     private BigDecimal totalRepaymentAmount;
 
-    private BigDecimal numberOfPayments;
+    private Integer numberOfPayments;
 
     private BigDecimal annualPercentageRate;
